Do not fail lunch when missing blobs
The lunch step will currently fail if you have not unzipped the blobs.
B2G build provides a script to download those blobs, but it relies on
the lunch step to complete properly. The script will fetch the EULA for
the matching blob set from Sony's website and expose the license to the
user who has to accept to trigger download, unzip and continue the build
process.

Signed-off-by: Alexandre Lissy <lissyx@lissyx.dyndns.org>